resolve_link handles HTML with embedded URI
When premailer was provided with a HTML string, resolve_link was choking
if the HTML contained a link on a single line. This is because
resolve_link and local_data? were both matching their URL regexp against
start of line instead of start of string to determine if the string was
a URL.

In general, \A provides the strictest match, and ^ is rarely needed.

Change-Id: Ib2ece04fa8dcf1f877c44ba39b7e9063c179185f
Reviewed-on: https://gerrit.causes.com/10477
Reviewed-by: Lann Martin <lann@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>