Correct Nokogiri-related warnings in non-breaking way
My prior PR may cause breaking changes to users of this library, and I
don't want to cause encoding-related problems for people. The documentation for
:input_encoding is clear enough to indicate that it should regularly be
set when using this library. 'ASCII-8BIT' is a sane default. This
patches the test to set the correct input encoding to silence the
Nokogiri warnings.