Support for Nokogiri Fragment (#348)
The integration with Nokogiri for parsing HTML currently requires and enforces a full HTML document (including doctype, html, and body tags). If you provide premailer an HTML fragment, Nokogiri will wrap the content unnecessarily.

This is happening because the Nokogiri integration is only using the HTML parser, not the HTML Fragment parser that Nokogiri provides.

While I augmented premailer to resolve a problem that I have on a project, this pull request proposes a solution for the problem and I would like to see it integrated into the official repository.