Add compatibility between rgba blending and preserve style attribute
If both the preserve style attribute and the rgba blending options are
enabled, there is a discrepancy between the style and the element
attribute because the style uses the original value instead of the
blended one.
This fix ensures that style and element attribute are outputted with the
same value.