Fix plaintext bug for links with single quote marks in the URI
The link regex was matching on either single or double quotes to detect the start/end of an href, so an href that includes a single quote mark as part if the link were not being captured correctly.

This tweaks the link substitution to treat links with single quotes independently from those with double quotes, so that the URIs such as https://en.wikipedia.org/wiki/King's_Gambit are handled correctly.

Closes #375, closes #376