Set "preserve_style_attribute" to true by default
Otherwise things may be overwritten when "css_to_attributes=true" and
"create_shorthands=true" (which are the defaults). For example given the css below:

    .my-class {
        background-repeat: no-repeat;
        background-position: bottom right;
        background-image: url("images/file.png");
    }

... when "background-image" is moved to HTML attribute and removed from the CSS, and
shorthand is created under "background" CSS attribute, background image becomes
overwritten. Having a "background" HTML attribute in that case doesn't help either. So
it's better to keep both even that's a duplication.