Change encoding of `html` string
`html` string is automatically encoded in utf-8 in modern Ruby versions.

This was causing an issue on the assumption of the input encoding by
Nokogiri adapter, causing double escaping of unicode `é`, `ç`, and `à`

This commit reverts the previous assertion, uses ASCII
representations of offending chars, and forces the input string to be
encoded in ISO-8859-1.

Close #430